Finish Challenge 16 and Set 2
This was probably the most fun one yet, as it felt a bit tricky, but you
had to solve it mosly on your own. I am really surprised this works, and
it shows a real attack angle. It was a bit unimpressive however that we
could not parse the decrypted as utf8...